pyproject.toml
[project]
name = "kindlepad"
version = "0.1.0"
description = "Kindle e-ink smart home control panel"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pillow>=11.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.28",
    "dirigera>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
